# specs/003-unified-interaction-check/contracts/interaction-check.openapi.yaml
openapi: 3.1.0
info:
  title: Unified interaction check
  version: 1.0.0
paths:
  /api/v1/interactions/check:
    post:
      summary: Tra cứu tương tác tổng hợp và lưu snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InteractionCheckRequest' }
      responses:
        '200':
          description: Kết quả trực tiếp, có thể chưa lưu được history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InteractionCheckResponse' }
  /api/v1/interaction-checks:
    get: { summary: Liệt kê snapshot của người dùng, responses: { '200': { description: Danh sách } } }
    delete: { summary: Xoá mọi snapshot của người dùng, responses: { '204': { description: Đã xoá } } }
  /api/v1/interaction-checks/{id}:
    get: { summary: Đọc snapshot, responses: { '200': { description: Chi tiết }, '404': { description: Không thuộc người dùng } } }
    delete: { summary: Xoá một snapshot, responses: { '204': { description: Đã xoá }, '404': { description: Không thuộc người dùng } } }
components:
  schemas:
    InteractionCheckRequest:
      type: object
      required: [drugIds, diseaseIds]
      properties:
        drugIds: { type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: { type: string, format: uuid } }
        diseaseIds: { type: array, maxItems: 20, uniqueItems: true, items: { type: string, format: uuid } }
    Citation:
      type: object
      required: [evidenceId, quote, source, sourceUrl]
      properties:
        evidenceId: { type: string }
        chunkId: { type: [string, 'null'], format: uuid }
        quote: { type: string, minLength: 1 }
        source: { type: string }
        sourceUrl: { type: string, format: uri }
    InteractionCheckResponse:
      type: object
      required: [checkId, historyStatus, severityScale, highlightId, items, notes, unavailable]
      properties:
        checkId: { type: [string, 'null'], format: uuid }
        historyStatus: { enum: [saved, not-saved] }
        severityScale: { type: array, items: { type: object } }
        highlightId: { type: [string, 'null'] }
        items: { type: array, items: { type: object } }
        notes: { type: array, items: { type: object } }
        unavailable: { type: array, items: { type: object } }